Added first tutorial (recipe) and supporting functions (in toolbox)
This first tutorial (which is a livescript for nicer presentation) shows how to:
(1) generate a Gabor scene and
(2) realize this scene on a particular display.

The tutoral computes several aspects of these scenes from first principles so as to show how to interact with ISETBio objects. The tutorial also visualizes different aspects of the display object.

Nicolas

Signed-off-by: Nicolas Cottaris <cottaris@sas.upenn.edu>